+++ internal/applications/remoter/pkg/config/config.go
@@ -1,6 +1,8 @@
 package config
 
 import (
+	"time"
+
 	"github.com/number571/hidden-lake/internal/applications/remoter/pkg/app/config"
 )
 
@@ -8,7 +10,7 @@ func GetConfigSettings(pCfg config.IConfig) SConfigSettings {
 	sett := pCfg.GetSettings()
 	return SConfigSettings{
 		SConfigSettings: config.SConfigSettings{
-			FExecTimeoutMS: sett.GetExecTimeoutMS(),
+			FExecTimeoutMS: uint64(sett.GetExecTimeout() / time.Millisecond),
 			FPassword:      sett.GetPassword(),
 		},
 	}