+++ internal/service/pkg/config/config.go
@@ -1,6 +1,8 @@
 package config
 
 import (
+	"time"
+
 	"github.com/number571/go-peer/pkg/client"
 	"github.com/number571/go-peer/pkg/encoding"
 	"github.com/number571/hidden-lake/internal/service/pkg/app/config"
@@ -10,11 +12,11 @@ func GetConfigSettings(pCfg config.IConfig, pClient client.IClient) SConfigSetti
 	sett := pCfg.GetSettings()
 	return SConfigSettings{
 		SConfigSettings: config.SConfigSettings{
+			FNetworkKey:       sett.GetNetworkKey(),
 			FMessageSizeBytes: sett.GetMessageSizeBytes(),
 			FWorkSizeBits:     sett.GetWorkSizeBits(),
-			FFetchTimeoutMS:   sett.GetFetchTimeoutMS(),
-			FQueuePeriodMS:    sett.GetQueuePeriodMS(),
-			FNetworkKey:       sett.GetNetworkKey(),
+			FFetchTimeoutMS:   uint64(sett.GetFetchTimeout() / time.Millisecond),
+			FQueuePeriodMS:    uint64(sett.GetQueuePeriod() / time.Millisecond),
 		},
 		// encoding.CSizeUint64 = payload64.Head()
 		FLimitMessageSizeBytes: pClient.GetPayloadLimit() - encoding.CSizeUint64,